requirements.txt
streamlit
snowflake-ml-python
matplotlib
seaborn